backend/requirements.txt
fastapi
uvicorn
google-genai
google-generativeai
langdetect
python-dotenv
python-multipart
motor
python-jose[cryptography]
bcrypt
# RAG dependencies
chromadb
sentence-transformers
datasets
tqdm
